Breaking NetJSON into files. Moving internally used methods by the serializer to NetJSON.Internals.SerializerUtilities(NetJSON.Internals.cs). Eliminating conditional instructions in NetJSON.cs.
